Summary: Porting to MS Windows Visual C++ 9.0 with CMake
Here I have gotten the port to MS Windows Visual C++ 9.0 to work with
CMake.  This required both changes to the CMake build system files and
to C++ code itself.

Here is what I did:

(*) CMakeLists.txt:

- Changed from 'BOOL' type to 'STRING' type for Trilinos_ENABLE_TESTS
and Trilinos_ENABLE_EXAMPLES.  This allows the ccmake and Windows
CMakeSetup GUI to leave these variables empty and maintain current
behavior.  Having them 'BOOL' caused the GUI to change them to OFF and
that messed up all of the dependency handling logic.  Now, the user
just needs to type 'ON' and 'OFF' to explicitly turn packages on and
off.

- Added option Trilinos_ENABLE_NATIVE_TEST_HARNESS_DEFAULT to turn
perl-based runtests testing support on and off.  The reason that I had
to do this is that CMake was creating some targets for these that
resulted in directory names that MS Visual Studio 9.0 2008 did not
like at all.  When WIN32 is defined, these are now off by default.
However, when WIN32 is not defined, they are on by default.

(*) TrilinosGlobalHelpers.cmake:

- Changed Trilinos_ENABLE_${PACKAGE},
${PACKAGE}_ENABLE_${OPTIONAL_DEP_PACKAGE}, ${PACKAGE}_ENABLE_TESTS,
and ${PACKAGE}_ENABLE_EXAMPLES from 'BOOL' to 'STRING' for the same
reason as described above.

(*) TrilinosPackageHelpers.cmake: Added conditional enable of runtests
targets according to Trilinos_ENABLE_NATIVE_TEST_HARNESS_DEFAULT as
described above.

(*) TrilinosProbeEnv.cmake: Changed from using 'uname -n' to the
portable SITE_NAME(...) macro to get the host name.  This is needed by
the HOST and XHOST options in the TRILINOS_ADD_TEST(...) macro.

(*) TrilinosCMakeEvaluation.tex: Few updates

(*) epetraext/CMakeLists.txt: Fixing bad option name

(*) packages/epetraext/test/CMakeLists.txt: Removing directories
SingletonFilter, View, and Overlap since they have no CMakeLists.txt
files and the MS Windows version of CMakeSetup did not like this and
gave nasty warning messages.

(*) Teuchos_BLAS_wrappers.hpp: Minor formatting change while I was
debugging.  Note however, that MS Visual C++ 9.0 is not too happy with
these BLAS functions that actually return std::complex<> objects.

(*) Teuchos_ConfigDefs.hpp, Teuchos_NullIteratorTraits.hpp: Added
logic to not set an iterator to null when _MSC_VER is defined.  I
don't know how to set an std::vector iterator to null on this
platform.

(*) Teuchos_ErrorPolling.cpp, Teuchos_ParameterList.hpp: Cleaning up
some warnings of unused catch objects.

(*) Teuchos_ScalarTraits.hpp: Put in static_cast<float>(...) to remove
some warnings about unsafe conversions.

(*) Teuchos_UnitTestRepository.cpp: Fixed a serious memory access
problem of indexing into an unsized std::string.  By default MS Visual
C++ 9.0 has the checked C++ library turned on and it caught this.
Thank goodness this was caught!

(*) Teuchos_XMLParser.cpp: Removing warnings about conversion problems
and reformated some.

(*) Array_UnitTest_helpers.hpp, Array_test.cpp,
teuchos/test/ParameterList/cxx_main.cpp, vector_UnitTests.cpp:
Removing more warnings about conversions and unused catch objects.

I tested this with MS Visual C++ 9.0 on my Windows laptop and also
with Cygwin/gcc on my same laptop.

The only test that fail with the cygwin/gcc build is:

    4 - Teuchos_Containers_test (SEGFAULT)

Note, however, that valgrind was complaining about this test on linux
(see bug 4181).

Two tests failed with the MS Visual C++ 9.0 build:

    80 - Triutils_read_test (Failed)
    83 - EpetraExt_inout_test (Failed)

The Triutils test segfaults.  The test EpetraExt_inoutP_test fails with lots of errors that look like:

  Epetra ERROR 3, /cygdrive/c/_mystuff/PROJECTS/Trilinos.base/Trilinos/packages/epetra/src/Epetra_CrsGraph.cpp, line 323
  Epetra ERROR 3, /cygdrive/c/_mystuff/PROJECTS/Trilinos.base/Trilinos/packages/epetra/src/Epetra_CrsMatrix.cpp, line 569
  Epetra ERROR 3, /cygdrive/c/_mystuff/PROJECTS/Trilinos.base/Trilinos/packages/epetra/src/Epetra_CrsMatrix.cpp, line 454
  ...

I am not sure what this is about but it could be something serious.

+++ packages/teuchos/src/Teuchos_ScalarTraits.hpp
@@ -311,7 +311,7 @@ struct ScalarTraits<float>
   }
   static inline float base()  {
 #ifdef HAVE_NUMERIC_LIMITS
-    return std::numeric_limits<float>::radix;
+    return static_cast<float>(std::numeric_limits<float>::radix);
 #else
     LAPACK<int, float> lp; return lp.LAMCH('B');
 #endif
@@ -325,7 +325,7 @@ struct ScalarTraits<float>
   }
   static inline float t()     {
 #ifdef HAVE_NUMERIC_LIMITS
-    return std::numeric_limits<float>::digits;
+    return static_cast<float>(std::numeric_limits<float>::digits);
 #else
     LAPACK<int, float> lp; return lp.LAMCH('N');
 #endif
@@ -339,7 +339,7 @@ struct ScalarTraits<float>
   }
   static inline float emin()  {
 #ifdef HAVE_NUMERIC_LIMITS
-    return std::numeric_limits<float>::min_exponent;
+    return static_cast<float>(std::numeric_limits<float>::min_exponent);
 #else
     LAPACK<int, float> lp; return lp.LAMCH('M');
 #endif
@@ -353,7 +353,7 @@ struct ScalarTraits<float>
   }
   static inline float emax()  {
 #ifdef HAVE_NUMERIC_LIMITS
-    return std::numeric_limits<float>::max_exponent;
+    return static_cast<float>(std::numeric_limits<float>::max_exponent);
 #else
     LAPACK<int, float> lp; return lp.LAMCH('L');
 #endif
@@ -386,9 +386,9 @@ struct ScalarTraits<float>
 #endif
   }
   static inline bool isnaninf(float x) { // RAB: 2004/05/28: Taken from NOX_StatusTest_FiniteValue.C
-    const float tol = 1e-6; // Any (bounded) number should do!
+    const float tol = 1e-6f; // Any (bounded) number should do!
     if( !(x <= tol) && !(x > tol) ) return true;                 // IEEE says this should fail for NaN
-    float z=0.0*x; if( !(z <= tol) && !(z > tol) ) return true;  // Use fact that Inf*0 = NaN
+    float z=0.0f*x; if( !(z <= tol) && !(z > tol) ) return true;  // Use fact that Inf*0 = NaN
     return false;
   }
   static inline void seedrandom(unsigned int s) {
